% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_top_level}
\alias{annotate_top_level}
\title{Annotate top level sections}
\usage{
annotate_top_level(file_contents)
}
\arguments{
\item{file_contents}{vector of file contents}
}
\value{
vector of annotations ("code", "yaml", or "text")
}
\description{
Annotate top level sections
}
