% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{extract_children}
\alias{extract_children}
\title{Summarise annotations}
\usage{
extract_children(
  annotations,
  file_contents,
  start = 1,
  end = length(file_contents)
)
}
\arguments{
\item{annotations}{vector of annotations}
}
\value{
tibble with start, end, and type columns
}
\description{
Summarise annotations
}
