% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_code.R
\name{extract_code}
\alias{extract_code}
\title{Extract all code from a parsed Quarto markdown doc}
\usage{
extract_code(x)
}
\arguments{
\item{x}{qmdparse_doc object}
}
\description{
Extract all code from a parsed Quarto markdown doc
}
