% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_headings}
\alias{extract_headings}
\title{Extract headings from a parsed Quarto markdown doc}
\usage{
extract_headings(x)
}
\arguments{
\item{x}{qmdparse_obj object}
}
\description{
Extract headings from a parsed Quarto markdown doc
}
