% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_named}
\alias{extract_named}
\title{Extract all elements from a parsed Quarto markdown doc with a matching name}
\usage{
extract_named(x, name)
}
\arguments{
\item{x}{qmdparse_obj object}

\item{name}{name of the element to extract}
}
\description{
Extract all elements from a parsed Quarto markdown doc with a matching name
}
