% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_text}
\alias{extract_text}
\title{Extract text from a parsed Quarto markdown doc}
\usage{
extract_text(x)
}
\arguments{
\item{x}{qmdparse_obj object}
}
\description{
Extract text from a parsed Quarto markdown doc
}
