% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_yaml}
\alias{extract_yaml}
\title{Extract yaml from a parsed Quarto markdown doc}
\usage{
extract_yaml(x)
}
\arguments{
\item{x}{qmdparse_obj object}
}
\description{
Extract yaml from a parsed Quarto markdown doc
}
