% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{print_tree}
\alias{print_tree}
\title{Print a tree of the parsed document}
\usage{
print_tree(obj, symbol = "", level = 0)
}
\arguments{
\item{obj}{The parsed document}

\item{symbol}{The symbol to use for the current node}

\item{level}{The current level of the tree}
}
\description{
This function prints a tree of the parsed document to the console.
}
