% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmd_map.R
\name{qmd_map}
\alias{qmd_map}
\title{Map a function to a qmd doc at a particular level}
\usage{
qmd_map(.path, .f, ..., .level = 0, .attribute = "cleaned_contents")
}
\arguments{
\item{.path}{Path to qmd doc}

\item{.f}{Function to map}

\item{...}{Additional arguments to pass to function}

\item{level}{Level to map function to}
}
\description{
Map a function to a qmd doc at a particular level
}
\examples{
qmd_map("tests/testthat/example_doc.qmd", stringi::str_count_words, .level = 1)
}
