% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmdparse-package.R
\docType{package}
\name{qmdparse-package}
\alias{qmdparse}
\alias{qmdparse-package}
\title{qmdparse: A simple package to parse qmd files.}
\description{
Parses `.qmd` files as simply as possible.
}
\author{
\strong{Maintainer}: Nic Crane \email{thisisnic@gmail.com}

}
\keyword{internal}
