% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_file_contents.R
\name{scan_contents}
\alias{scan_contents}
\title{Annotate top level sections}
\usage{
scan_contents(file_contents, level, offset)
}
\arguments{
\item{file_contents}{vector of file contents}

\item{level}{depth within doc}

\item{offset}{offset of where this section starts}
}
\value{
vector of annotations ("code", "yaml", or "text")
}
\description{
Annotate top level sections
}
