% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{parse_contents}
\alias{parse_contents}
\title{Annotate top level sections}
\usage{
parse_contents(file_contents, offset = 0)
}
\arguments{
\item{file_contents}{vector of file contents}
}
\value{
vector of annotations ("code", "yaml", or "text")
}
\description{
Annotate top level sections
}
